+++ RevitAddin.DI.Simple/Services/Commands/ICommand.cs
@@ -1,0 +1,7 @@
+﻿namespace RevitAddin.DI.Simple.Services.Commands
+{
+    public interface ICommand
+    {
+        void Execute();
+    }
+}