+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [1.0.0] / 2023-12-06
+- Add `RevitCommand` with `ICommand` interface.
 - Add `MyService` sample
 - Add `RevitCommandService`
 - Add `Services`